changes for double in SettingForApplication

+++ Source/SSCore/src/main/java/com/realtech/socialsurvey/core/services/settingsmanagement/SettingsSetter.java
@@ -84,5 +84,5 @@ public interface SettingsSetter {
 	 * @return
 	 * @throws InvalidSettingsStateException
 	 */
-	public OrganizationUnit getLowestSetterLevel(int setUnitValue) throws InvalidSettingsStateException;
+	public OrganizationUnit getLowestSetterLevel( double setUnitValue) throws InvalidSettingsStateException;
 }
